+++ packages/shared-components/src/formio/components/core/maalgruppe/Maalgruppe.form.ts
